api/src/app/requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
uvloop==0.16.0
orjson==3.6.7
PyJWT==2.4.0
backoff==2.0.1
SQLAlchemy==1.4.37
psycopg2-binary==2.9.3
pydantic==1.9.1
python-dotenv==0.20.0
sqlmodel==0.0.6
asyncpg==0.25.0
aio-pika==8.0.3